+++ pkg/deployment/resources/inspector/pvs_gvk.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -21,24 +21,23 @@
 package inspector
 
 import (
-	"time"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/throttle"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/constants"
 )
 
-func NewDefaultThrottle() throttle.Components {
-	return throttle.NewThrottleComponents(
-		30*time.Second, // ArangoDeploymentSynchronization
-		30*time.Second, // ArangoMember
-		30*time.Second, // ArangoTask
-		30*time.Second, // Node
-		30*time.Second, // PV
-		15*time.Second, // PVC
-		time.Second,    // Pod
-		30*time.Second, // PDB
-		10*time.Second, // Secret
-		10*time.Second, // Service
-		30*time.Second, // SA
-		30*time.Second, // ServiceMonitor
-		15*time.Second) // Endpoints
+func (p *persistentVolumesInspectorV1) GroupVersionKind() schema.GroupVersionKind {
+	return constants.PersistentVolumeGKv1()
+}
+
+func (p *persistentVolumesInspectorV1) GroupVersionResource() schema.GroupVersionResource {
+	return constants.PersistentVolumeGRv1()
+}
+
+func (p *persistentVolumesInspector) GroupKind() schema.GroupKind {
+	return constants.PersistentVolumeGK()
+}
+
+func (p *persistentVolumesInspector) GroupResource() schema.GroupResource {
+	return constants.PersistentVolumeGR()
 }